+++ foxes/models/turbine_models/__init__.py
@@ -2,4 +2,4 @@ from .set_XYHD import SetXYHD
 from .kTI_model import kTI
 from .set_farm_vars import SetFarmVars
 from .thrust2ct import Thrust2Ct
-from .max_power import MaxPower
+from .power_mask import PowerMask